@@ -15,7 +15,7 @@
 	"devDependencies": {
 		"@lucide/svelte": "^0.503.0",
 		"@sveltejs/adapter-node": "^5.2.9",
-		"@sveltejs/kit": "^2.0.0",
+		"@sveltejs/kit": "^2.5.27",
 		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"autoprefixer": "^10.4.20",
 		"bits-ui": "^1.3.19",
@@ -29,7 +29,7 @@
 		"tailwind-variants": "^1.0.0",
 		"tailwindcss": "^3.4.9",
 		"tailwindcss-animate": "^1.0.7",
-		"typescript": "^5.0.0",
+		"typescript": "^5.5.0",
 		"vite": "^5.4.11"
 	},
 	"dependencies": {
